+++ src/web/utils/ErrorView.tsx
@@ -1,0 +1,30 @@
+const analyticsIdentifier = (errorId?: string, errorHash?: string) => {
+  if (!errorHash && !errorId) return 'No error identifier';
+  console.log('errorHash', errorHash);
+  console.log('errorId', errorId);
+  return errorHash
+    ? `error hash: ${errorHash}` // From backend
+    : `error ID: ${errorId}`; // Generated by frontend to make it possible to triage user reported errors.
+};
+
+export function ErrorView({
+  message,
+  errorId,
+  errorHash,
+}: {
+  message?: string;
+  errorId?: string;
+  errorHash?: string;
+}) {
+  return (
+    <div className='error-content'>
+      <img alt='Error icon' src='/uid2-logo.png' />
+      <div>Error</div>
+      <div>
+        {message ??
+          'Unexpected error encountered, please contact support if the problem persists and provide the information below'}
+      </div>
+      <div>({analyticsIdentifier(errorId, errorHash)})</div>
+    </div>
+  );
+}
