+++ src/web/utils/PortalErrorBoundary.tsx
@@ -1,15 +1,9 @@
-import { errorHandler, ErrorView, RenderedErrorProps } from './errorHandler';
+import { errorHandler, RenderedErrorProps } from './errorHandler';
+import { ErrorView } from './ErrorView';
 
 function PortalErrorComponent(props: RenderedErrorProps) {
   const { errorId, errorHash } = props;
-  return (
-    <ErrorView
-      errorId={errorId}
-      errorHash={errorHash}
-      // Override whatever message we might have, messages we might see are 99% unhelpful
-      message='Unexpected error encountered, please contact support if the problem persists and provide the information below'
-    />
-  );
+  return <ErrorView errorId={errorId} errorHash={errorHash} />;
 }
 
 export const PortalErrorBoundary = errorHandler(PortalErrorComponent);
