+++ src/web/utils/errorHandler.tsx
@@ -22,31 +22,6 @@ export interface RenderedErrorProps {
   componentStack?: string;
 }
 
-export function analyticsIdentifier(errorId?: string, errorHash?: string) {
-  if (!errorHash && !errorId) return 'No error identifier';
-  return errorHash
-    ? `error hash: ${errorHash}` // From backend
-    : `error ID: ${errorId}`; // Generated by frontend to make it possible to triage user reported errors.
-}
-
-export function ErrorView({
-  message,
-  errorId,
-  errorHash,
-}: {
-  message: string;
-  errorId?: string;
-  errorHash?: string;
-}) {
-  return (
-    <div className='error-content'>
-      <img alt='Error icon' src='/uid2-logo.png' />
-      <div>Error</div>
-      <div>{message}</div>
-      <div>({analyticsIdentifier(errorId, errorHash)})</div>
-    </div>
-  );
-}
 export interface ErrorBoundaryPropType {
   children: ReactNode;
   [key: string]: unknown;
